+++ Migrations/ApplicationDbContextModelSnapshot.cs
@@ -77,8 +77,8 @@ namespace BookingEngine.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("PhoneNumber")
-                        .HasColumnType("int");
+                    b.Property<long>("PhoneNumber")
+                        .HasColumnType("bigint");
 
                     b.Property<int>("ZipCode")
                         .HasColumnType("int");
